requirements.txt
openai>=1.0.0
anthropic>=0.20.0
chromadb>=0.4.0
langgraph>=0.1.0
pyautogen>=0.2.0
asyncio
